+++ examples/replication_complex.php
@@ -56,10 +56,9 @@ $options = [
     'commands' => [
         'hmgetall' => 'HashMultipleGetAll',
     ],
-    'replication' => function () {
+    'replication' => static function () {
         $strategy = new ReplicationStrategy();
         $strategy->setScriptReadOnly(HashMultipleGetAll::BODY);
-
         $replication = new MasterSlaveReplication($strategy);
 
         return $replication;