+++ src/Command/Redis/VLINKS.php
@@ -58,7 +58,7 @@ class VLINKS extends RedisCommand
             if ($this->withScores) {
                 foreach ($data as $key => $value) {
                     if ($value === array_values($value)) {
-                        $data[$key] = CommandUtility::arrayToDictionary($value, function ($key, $value) {
+                        $data[$key] = CommandUtility::arrayToDictionary($value, static function ($key, $value) {
                             return [$key, (float) $value];
                         });
                     } else {