+++ tests/Predis/Retry/RetryTest.php
@@ -39,11 +39,10 @@ class RetryTest extends TestCase
         $retry = new Retry($backoffStrategy, $retries);
         $retriesCount = 0;
 
-        $callable = function () use (&$retriesCount, $retries) {
+        $callable = static function () use (&$retriesCount, $retries) {
             if ($retriesCount >= $retries) {
                 return;
             }
-
             ++$retriesCount;
             throw new StreamInitException();
         };
@@ -82,7 +81,7 @@ class RetryTest extends TestCase
             throw new StreamInitException();
         };
 
-        $failCallable = function () use (&$retriesCount) {
+        $failCallable = static function () use (&$retriesCount) {
             ++$retriesCount;
         };