+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Unreleased
 - Fixed Sentinel getParameters() executed on string configuration (#1649)
 - Fixed Sentinel discovery methods not catching `StreamInitException` on connection failure (#1650)
+- Make callables `static` as much as possible, and enforce code style using php-cs-fixer. (#1651)
 
 ## v3.4.1 (2026-02-23)
 ### Added